Primary content nodes are *children* of `<body>` in content documents which also match the `@id` of a target element in the decor.
Previously:
 all child-nodes were removed from cached content docs when paging in -
content nodes were placed in the view,
other nodes were dropped into the aether.
When paging out, content nodes were restored to the cached content doc.

After this commit:
 only content nodes are removed from the cached content doc,
and they are replaced with placeholder nodes so that paging out can restore
the originals.

This change makes it possible to save original content
(by copying it to another node) and transform it before paging in.